Avoid the segfault reported in parrot's TT #10
The original segfault was caused because of the recent change to parrot's
exception handling (which we never addressed due to other build issues); When
branching to the exception handler label after an exception is invoked, there
is no more implicit "catch". You need to explicitly remove the exception
handler or it can be invoked again, causing hijinks to ensure.

This only gets us further through make test; cmd_expr.t still hangs.



git-svn-id: http://partcl.googlecode.com/svn/trunk@189 6cb8db7d-f34b-0410-8f57-4f83c6281724